Use feature folder structuring, use ReduxDucks, Add minor styling and changes, Add Add a new book section
