--- src/components/BookList.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import Book from './Book';
-import './styles/booklist.css';
-
-const BookList = () => (
-  <section className="book-list-container">
-    <Book />
-  </section>
-);
-
-export default BookList;
