+++ src/components/BooksContainer/BookList.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import Book from './BookList/Book';
+import '../styles/booklist.css';
+
+const BookList = () => (
+  <Book />
+);
+
+export default BookList;
