+++ src/components/BooksContainer/BookList/Book.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './styles/book.css';
+import '../../styles/book.css';
 
 const temp = [
   {
