+++ src/components/Pages/Categories.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Categories = () => (
+  <h1>
+    UNDER CONSTRUCTION
+  </h1>
+);
+
+export default Categories;
