+++ src/components/styles/booklist.css
@@ -11,8 +11,7 @@ main {
 }
 
 .main-container {
-  width: 90%;
-  border: 2px solid black;
+  width: 98%;
   display: flex;
   flex-direction: column;
 }
