+++ src/components/styles/navbar.css
@@ -4,9 +4,14 @@
   left: 0;
   justify-content: space-between;
   display: flex;
-  align-items: baseline;
+  align-items: center;
   width: 100%;
   background-color: white;
+}
+
+.website-logo-name {
+  margin: 15px 0;
+  color: rgb(3, 118, 196);
 }
 
 .navbar > .left-side {
@@ -21,6 +26,7 @@
 .left-side > ul {
   list-style-type: none;
   display: flex;
+  margin: 0;
 }
 
 .left-side > ul > li {
@@ -28,5 +34,18 @@
 }
 
 .right-side {
-  margin-right: 8%;
+  margin: 7px 8% 0 0;
+}
+
+.right-side > img {
+  width: 70%;
+}
+
+.links {
+  text-decoration: none;
+  color: rgb(136, 136, 136);
+}
+
+.activelinks {
+  color: rgb(71, 71, 71);
 }
